moved interfaces under the lib directory
